Moved API search results back to the HomeFragment.java file.
