+++ app/src/main/java/com/example/whereto/fragments/HomeFragment.java
@@ -13,10 +13,14 @@ import com.example.whereto.Constants;
 import com.example.whereto.Question;
 import com.example.whereto.QuestionsAdapter;
 import com.example.whereto.R;
+import com.example.whereto.models.YelpService;
 import com.example.whereto.models.YelpServiceInterface;
 
 import java.util.ArrayList;
 
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
@@ -39,6 +43,45 @@ public class HomeFragment extends Fragment {
     public void getYelpData() {
         final Retrofit retrofit = new Retrofit.Builder().baseUrl(Constants.BASE_URL).addConverterFactory(GsonConverterFactory.create()).build();
         final YelpServiceInterface yelpServiceInterface = retrofit.create(YelpServiceInterface.class);
-        yelpServiceInterface.getSearchResults();
+        getSearchResults(yelpServiceInterface);
+    }
+
+    public void getSearchResults(YelpServiceInterface yelpServiceInterface) {
+        String searchTerm = "Avocado Toast"; // TODO: replace with user input
+        String searchCategory = "Concert"; // TODO: replace with user input
+        String location = "San Francisco"; // TODO: replace with user input
+
+        getRestaurantResults(yelpServiceInterface, searchTerm, location);
+        getEventsResults(yelpServiceInterface, searchCategory, location);
+    }
+
+    public void getRestaurantResults(final YelpServiceInterface yelpServiceInterface, final String searchTerm, final String location) {
+        Call<YelpService> restaurantCall = yelpServiceInterface.searchRestaurants("Bearer "+ Constants.API_KEY, searchTerm, location);
+        restaurantCall.enqueue(new Callback<YelpService>() {
+            @Override
+            public void onResponse(Call<YelpService> call, Response<YelpService> response) {
+                // TODO: store and return response based on the user's preferences
+            }
+
+            @Override
+            public void onFailure(Call<YelpService> call, Throwable t) {
+                t.printStackTrace();
+            }
+        });
+    }
+
+    public void getEventsResults(final YelpServiceInterface yelpServiceInterface, final String searchCategory, final String location) {
+        Call<YelpService> eventsCall = yelpServiceInterface.searchEvents("Bearer "+ Constants.API_KEY, searchCategory, location);
+        eventsCall.enqueue(new Callback<YelpService>() {
+            @Override
+            public void onResponse(Call<YelpService> eventsCall, Response<YelpService> response) {
+                // TODO: store and return response based on the user's preferences
+            }
+
+            @Override
+            public void onFailure(Call<YelpService> eventsCall, Throwable t) {
+                t.printStackTrace();
+            }
+        });
     }
 }
